+++ Internal/Utils/Utils.cs
@@ -194,5 +194,68 @@ namespace Anatawa12.AvatarOptimizer
 
         public static EqualsHashSet<T> ToEqualsHashSet<T>(this HashSet<T> hashSet) =>
             new EqualsHashSet<T>(hashSet);
+
+        public static bool IsPowerOfTwo(this int x) => x != 0 && (x & (x - 1)) == 0;
+
+        public static float MinPowerOfTwoGreaterThan(float x)
+        {
+            if (x <= 0) throw new ArgumentOutOfRangeException(nameof(x), x, "x must be positive");
+
+            if (x < 1)
+            {
+                var r = 1f;
+                while (r / 2 > x) r /= 2;
+                return r;
+            }
+            else
+            {
+                var r = 1f;
+                while (r < x) r *= 2;
+                return r;
+            }
+        }
+
+        public static int LeastCommonMultiple(params int[] numbers) => numbers.Length == 0 ? 0 : numbers.Aggregate(LeastCommonMultiple);
+
+        public static int LeastCommonMultiple(int a, int b)
+        {
+            if (a == 0 || b == 0) return 0;
+            return Math.Abs(a * b) / GreatestCommonDivisor(a, b);
+        }
+
+        public static int GreatestCommonDivisor(int a, int b)
+        {
+            if (a == 0) return b;
+            if (b == 0) return a;
+            while (b != 0)
+            {
+                var t = b;
+                b = a % b;
+                a = t;
+            }
+
+            return a;
+        }
+
+        public static int MostSignificantBit(int x) => MostSignificantBit((uint)x);
+
+        public static int MostSignificantBit(uint x)
+        {
+            // https://github.com/microsoft/mimalloc/blob/fab7329c7a3eee64e455e0a7aea7566eb2038cf3/src/page-queue.c#L67-L80
+            
+            // de Bruijn multiplication, see <http://supertech.csail.mit.edu/papers/debruijn.pdf>
+            ReadOnlySpan<byte> debruijn = new byte[]{
+                31,  0, 22,  1, 28, 23, 18,  2, 29, 26, 24, 10, 19,  7,  3, 12,
+                30, 21, 27, 17, 25,  9,  6, 11, 20, 16,  8,  5, 15,  4, 14, 13,
+            };
+            
+            x |= x >> 1;
+            x |= x >> 2;
+            x |= x >> 4;
+            x |= x >> 8;
+            x |= x >> 16;
+            x++;
+            return debruijn[(int)((x * 0x076be629) >> 27)];
+        }
     }
 }